Detect user's default shell
