+++ jupyter_forward/console.py
@@ -1,0 +1,5 @@
+from __future__ import annotations
+
+import rich.console
+
+console = rich.console.Console()
